automation/requirements.txt
requests
beautifulsoup4
pillow
rembg
cloudinary
python-dotenv
supabase
